fix passing environment variables
fixes https://github.com/containers/podman-compose/issues/1011

The problem before this fix:
if the environment section contains assignments without value, such as:
environment:
- SOME_VAR

it will get added to the podman.args as is, just "-e SOME_VAR" which is likely not intended - at least from what i can tell, in these situations vars should be looked up in the environment added by, for example .env files. 

The fix:
if an entry in the environment section doesn't contain a value, we look it up in compose.environ and add it if found.